fix(sampler): cast the sampled ids on every gather-based logprob path
`torch.gather` only takes int64 indices, and the sampling ops' width is
whatever the compiler picked. `gather_logprobs` already re-cast; the
generative-scoring path did not, so a request carrying logprob_token_ids
died inside `gather_specific_token_logprobs`. Sync, async and optimum all
reach it.

+++ vllm_rbln/v1/sample/rbln_sampler.py
@@ -312,7 +312,7 @@ class RBLNSampler(VLLMSampler):
         if sampling_metadata.logprob_token_ids:
             assert raw_logprobs is not None
             logprob_token_ids_tensors = self.gather_specific_token_logprobs(
-                raw_logprobs, sampling_metadata.logprob_token_ids, sampled
+                raw_logprobs, sampling_metadata.logprob_token_ids, sampled.long()
             )
 
         if num_logprobs is None:
